requirement.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
mysqlclient==2.1.0
PyMySQL==1.0.2
sqlclient==21.4.24
django-rest-auth==0.9.5
requests==2.27.1